apached test

+++ apache/build.sh
@@ -2,6 +2,15 @@
   set -x
   source ../home_var.sh
 
+  if [ $# != 0 ] && [ ! -n "${lib_with_path_map[$1]}" ]; then
+    echo "unknown allocators: $1"
+    exit 1
+  fi
+
+  if [ $# != 0 ]; then
+    export LD_LIBRARY_PATH=${preload_map[$1]}
+  fi
+
   rm -rf 'httpd-2.4.35'
 
   tar -xvf httpd-2.4.35.tar.gz
@@ -15,7 +24,7 @@
  # cp ../$config_vars ./build/config_vars.mk
   if [ $# != 0 ] && [ "${lib_with_path_map[$1]}" != "NULL" ]; then
     mv ./build/config_vars.mk ./build/config_vars.mk.bk
-    cat ./build/config_vars.mk.bk | sed "s;\-pthread;-rdynamic ${lib_with_path_map[$1]} -pthread;" | sed "s;\-lpthread;-rdynamic ${lib_with_path_map[$1]} -lpthread;" > ./build/config_vars.mk
+    cat ./build/config_vars.mk.bk | sed "s;\-pthread;-rdynamic ${lib_with_path_map[$1]} -pthread;" | sed "s;\-lpthread;-rdynamic ${lib_with_path_map[$1]} -lpthread;" | sed 's/^CPPFLAGS =/CPPFLAGS = -Wl,--no-as-needed/' | sed 's/^CFLAGS =/CFLAGS = -Wl,--no-as-needed/' > ./build/config_vars.mk
   fi
   make
   make install
